feat(payment): implement persistence adapter with JPA repository and domain mapping

+++ src/main/java/com/microservices/payment/shared/domain/model/PageDomain.java
@@ -6,5 +6,5 @@ public record PageDomain <T>(
         List<T> content,
         int currentPage,
         int totalPages,
-        int totalElements
+        long totalElements
 ){}
